+++ packages/graz/src/types/wallet.ts
@@ -12,6 +12,8 @@ export enum WalletType {
   WC_LEAP_MOBILE = "wc_leap_mobile",
   // eslint-disable-next-line @typescript-eslint/naming-convention
   WC_COSMOSTATION_MOBILE = "wc_cosmostation_mobile",
+  // eslint-disable-next-line @typescript-eslint/naming-convention
+  METAMASK_SNAP_LEAP = "metamask_snap_leap",
 }
 
 export const WALLET_TYPES = [
@@ -23,6 +25,7 @@ export const WALLET_TYPES = [
   WalletType.WC_KEPLR_MOBILE,
   WalletType.WC_LEAP_MOBILE,
   WalletType.WC_COSMOSTATION_MOBILE,
+  WalletType.METAMASK_SNAP_LEAP,
 ];
 
 export type Wallet = Pick<
@@ -39,3 +42,6 @@ export type Wallet = Pick<
   subscription?: (reconnect: () => void) => void;
   init?: () => Promise<unknown>;
 };
+
+export type SignDirectParams = Parameters<Wallet["signDirect"]>;
+export type SignAminoParams = Parameters<Wallet["signAmino"]>;