+++ templates/default/src/ui/wallet/connect-button.tsx
@@ -71,6 +71,9 @@ export const WalletConnectButton = () => {
             {wallets.wc_cosmostation_mobile ? (
               <Button onClick={() => handleConnect(WalletType.WC_COSMOSTATION_MOBILE)}>Cosmostation Mobile</Button>
             ) : null}
+            {wallets.metamask_snap_leap ? (
+              <Button onClick={() => handleConnect(WalletType.METAMASK_SNAP_LEAP)}>Metamask Snap Leap</Button>
+            ) : null}
           </Stack>
         </Modal>
       )}